+++ src/internal/form-control.ts
@@ -1,36 +1,38 @@
 import { html, TemplateResult } from 'lit';
 import { classMap } from 'lit-html/directives/class-map';
+import { ifDefined } from 'lit-html/directives/if-defined';
 
-export interface FormControlProps {
-  /** The input id, used to map the input to the label */
-  inputId: string;
+export const renderFormControl = (
+  props: {
+    /** The input id, used to map the input to the label */
+    inputId: string;
 
-  /** The size of the form control */
-  size: 'small' | 'medium' | 'large';
+    /** The size of the form control */
+    size: 'small' | 'medium' | 'large';
 
-  /** The label id, used to map the label to the input */
-  labelId?: string;
+    /** The label id, used to map the label to the input */
+    labelId?: string;
 
-  /** The label text (if the label slot isn't used) */
-  label?: string;
+    /** The label text (if the label slot isn't used) */
+    label?: string;
 
-  /** Whether or not a label slot has been provided. */
-  hasLabelSlot?: boolean;
+    /** Whether or not a label slot has been provided. */
+    hasLabelSlot?: boolean;
 
-  /** The help text id, used to map the input to the help text */
-  helpTextId?: string;
+    /** The help text id, used to map the input to the help text */
+    helpTextId?: string;
 
-  /** The help text (if the help-text slot isn't used) */
-  helpText?: string;
+    /** The help text (if the help-text slot isn't used) */
+    helpText?: string;
 
-  /** Whether or not a help text slot has been provided. */
-  hasHelpTextSlot?: boolean;
+    /** Whether or not a help text slot has been provided. */
+    hasHelpTextSlot?: boolean;
 
-  /** A function that gets called when the label is clicked. */
-  onLabelClick?: (event: MouseEvent) => void;
-}
-
-export const renderFormControl = (props: FormControlProps, input: TemplateResult) => {
+    /** A function that gets called when the label is clicked. */
+    onLabelClick?: (event: MouseEvent) => void;
+  },
+  input: TemplateResult
+) => {
   const hasLabel = props.label ? true : !!props.hasLabelSlot;
   const hasHelpText = props.helpText ? true : !!props.hasHelpTextSlot;
 
@@ -48,7 +50,7 @@ export const renderFormControl = (props: FormControlProps, input: TemplateResult
     >
       <label
         part="label"
-        id=${props.labelId}
+        id=${ifDefined(props.labelId)}
         class="form-control__label"
         for=${props.inputId}
         aria-hidden=${hasLabel ? 'false' : 'true'}
@@ -61,7 +63,7 @@ export const renderFormControl = (props: FormControlProps, input: TemplateResult
 
       <div
         part="help-text"
-        id=${props.helpTextId}
+        id=${ifDefined(props.helpTextId)}
         class="form-control__help-text"
         aria-hidden=${hasHelpText ? 'false' : 'true'}
       >
@@ -70,3 +72,30 @@ export const renderFormControl = (props: FormControlProps, input: TemplateResult
     </div>
   `;
 };
+
+export function getLabelledBy(props: {
+  /** The label id, used to map the label to the input */
+  labelId: string;
+
+  /** The label text (if the label slot isn't used) */
+  label: string;
+
+  /** Whether or not a label slot has been provided. */
+  hasLabelSlot: boolean;
+
+  /** The help text id, used to map the input to the help text */
+  helpTextId: string;
+
+  /** The help text (if the help-text slot isn't used) */
+  helpText: string;
+
+  /** Whether or not a help text slot has been provided. */
+  hasHelpTextSlot: boolean;
+}) {
+  const labelledBy = [
+    props.label || props.hasLabelSlot ? props.labelId : '',
+    props.helpText || props.hasHelpTextSlot ? props.helpTextId : ''
+  ].filter(val => val);
+
+  return labelledBy.join(' ') || undefined;
+}