+++ docs/resources/changelog.md
@@ -11,6 +11,7 @@ _During the beta period, these restrictions may be relaxed in the event of a mis
 - Added `click()` method to `sl-checkbox`, `sl-radio`, and `sl-switch`
 - Fixed a bug in `sl-tooltip` where events weren't properly cleaned up on disconnect
 - Fixed a bug in `sl-tooltip` where they wouldn't display after toggling `disabled` off and on again [#391](https://github.com/shoelace-style/shoelace/issues/391)
+- Fixed incorrect event names for `sl-after-show` and `sl-after-hide` in `sl-details`
 - Improved a11y for disabled buttons that are rendered as links
 - Improved a11y for `sl-button-group`