+++ docs/resources/changelog.md
@@ -15,7 +15,8 @@ _During the beta period, these restrictions may be relaxed in the event of a mis
 - Fixed a bug in `sl-details` where `show()` and `hide()` would toggle the control when disabled
 - Fixed incorrect event names for `sl-after-show` and `sl-after-hide` in `sl-details`
 - Improved a11y for disabled buttons that are rendered as links
-- Improved a11y for `sl-button-group`
+- Improved a11y for `sl-button-group` by adding the correct `role` attribute
+- Improved a11y for `sl-input`, `sl-range`, `sl-select`, and `sl-textarea` so labels and helper text are read properly by screen readers
 - Removed `sl-show`, `sl-hide`, `sl-after-show`, `sl-after-hide` events from `sl-color-picker` (the color picker's visibility cannot be controlled programmatically so these shouldn't have been exposed; the dropdown events now bubble up so you can listen for those instead)
 - Reworked `sl-button-group` so it doesn't require light DOM styles